(refactor) e2e: Centralize types and improve type usage
This PR centralizes type annotations used in e2e tests and improves type usage. Key changes include:

- Adding the Playwright ESLint plugin to enforce best practices in e2e tests.
- Moving test-specific types to a dedicated types file and importing them where needed.
- Using consistent type imports in e2e tests.
- Cleaning up eslint overrides for various test file types.

With these changes, e2e tests should be more consistent and easier to maintain.
