+++ package.json
@@ -55,6 +55,7 @@
     "eslint": "^8.55.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jest-dom": "^5.4.0",
+    "eslint-plugin-playwright": "^2.1.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-testing-library": "^6.2.2",
     "husky": "^8.0.3",
